+++ src/gii/Generator.php
@@ -56,7 +56,7 @@ class Generator extends \yii\gii\Generator
         parent::init();
 
         if ($this->migrationName === null) {
-            $this->migrationName = 'm' . gmdate('ymd_His') . '_add_email_template';
+            $this->migrationName = 'm' . \gmdate('ymd_His') . '_add_email_template';
         }
 
         $this->hint = 'All tokens wrapped in {} will be replaced by real data';
@@ -83,7 +83,7 @@ class Generator extends \yii\gii\Generator
      */
     public function rules()
     {
-        return array_merge(parent::rules(), [
+        return \array_merge(parent::rules(), [
             [['migrationName'], 'safe'],
             [['key', 'subject', 'body'], 'required'],
             [['hint'], 'string', 'max' => 500],
@@ -95,7 +95,7 @@ class Generator extends \yii\gii\Generator
      */
     public function attributeLabels()
     {
-        return array_merge(parent::attributeLabels(), [
+        return \array_merge(parent::attributeLabels(), [
             'key'       => 'Template key',
             'subject'   => 'Letter subject',
             'body'      => 'Letter body',